+++ bin/build.sh
@@ -10,6 +10,6 @@ host="${FLUX_PUBLISH_DOCKER_HOST:=}"
 host_with_slash="${host}${host:+/}"
 user="${FLUX_PUBLISH_DOCKER_USER:=fluxfw}"
 image="$host_with_slash$user/$name"
-tag="`$bin/PATH/host/get-release-tag "$root"`"
+tag="`$bin/HOST_PATH/get-release-tag "$root"`"
 
 docker build "$root" --pull -t "$image:$tag" -t "$image:latest"