+++ bin/get-release-tag.sh
@@ -4,4 +4,4 @@ set -e
 
 bin="`dirname "$(realpath "$0")"`"
 
-echo -n "v`$bin/PATH/host/get-release-version "$@"`"
+echo -n "v`$bin/HOST_PATH/get-release-version "$@"`"